functionalComponent is deprecated
