+++ kotlin-react/src/main/kotlin/react/FunctionComponent.kt
@@ -1,3 +1,5 @@
+@file:Suppress("NOTHING_TO_INLINE")
+
 package react
 
 external interface FunctionComponent<in P : RProps> :
@@ -5,8 +7,29 @@ external interface FunctionComponent<in P : RProps> :
 
 typealias FC<P> = FunctionComponent<P>
 
-@Deprecated(
-    message = "Legacy type alias",
-    ReplaceWith("FunctionComponent", "react.FunctionComponent")
-)
-typealias FunctionalComponent<P> = FunctionComponent<P>
+/**
+ * Get function component from [func]
+ */
+fun <P : RProps> functionComponent(
+    displayName: String? = null,
+    func: RBuilder.(props: P) -> Unit,
+): FC<P> {
+    val fc: dynamic = { props: P ->
+        buildElements {
+            func(props)
+        }
+    }
+    if (displayName != null) {
+        fc.displayName = displayName
+    }
+    return fc.unsafeCast<FunctionComponent<P>>()
+}
+
+/**
+ * Get function component from [func]
+ */
+inline fun <P : RProps> fc(
+    displayName: String? = null,
+    noinline func: RBuilder.(props: P) -> Unit,
+) =
+    functionComponent(displayName, func)
